fix github issues/2931 解决结果集大于 Integer 异常问题

+++ mybatis-plus-core/src/main/java/com/baomidou/mybatisplus/core/assist/ISqlRunner.java
@@ -51,7 +51,7 @@ public interface ISqlRunner {
 
     Object selectObj(String sql, Object... args);
 
-    int selectCount(String sql, Object... args);
+    long selectCount(String sql, Object... args);
 
     Map<String, Object> selectOne(String sql, Object... args);
